+++ NEWS.md
@@ -10,6 +10,8 @@
 
 * [libmdl] Change include directory from `$PREFIX/include` to `$PREFIX/include/mdl`
 * [mdl] Implement Title3 for `NcursesDisplayDriver`
+* [mdl] Implement Title3 for `HtmlDisplayDriver`
+* [mdl] Implement Title3 for `LaTeXDisplayDriver`
 
 ## [v1.0.0] 2016-07-15